Cleaned up code after changing to PyNumero
